+++ package.json
@@ -11,12 +11,14 @@
     "@types/node": "^12.7.8",
     "@types/react": "^16.9.3",
     "@types/react-dom": "^16.9.1",
+    "@types/react-transition-group": "^4.2.3",
     "@types/redux": "^3.6.0",
     "@types/styled-components": "^4.1.19",
     "axios": "^0.19.0",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-scripts": "3.1.2",
+    "react-transition-group": "^4.3.0",
     "redux": "^4.0.4",
     "styled-components": "^4.4.0",
     "styled-reset": "^4.0.1",
