+++ src/components/VideoDetail/styles.css
@@ -1,0 +1,31 @@
+.fade {
+  background-color: red;
+  color: #fff;
+  height: 0;
+  overflow: hidden;
+}
+@keyframes fadeIn {
+  0% {
+    height: 0;
+  }
+
+  50% {
+    height: 50%;
+  }
+  100% {
+    height: 100%;
+  }
+}
+
+@keyframes fadeOut {
+  0% {
+    height: 100%;
+  }
+
+  50% {
+    height: 50%;
+  }
+  100% {
+    height: 0;
+  }
+}
