+++ src/components/VideoList/VideoList.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import AppContext from "../../contexts/AppContext";
 
 import styled from "styled-components";
-import Grid from "@material-ui/core/Grid";
+import { Grid } from "@material-ui/core";
 import Video from "../Video/Video";
 import VideoDetail from "../VideoDetail/VideoDetail";
 
@@ -41,8 +41,7 @@ const WrapVideoDetailGrid = styled(Grid)`
   && {
     /* transition: 0.5s; */
   }
-`
-
+`;
 
 const VideoList: React.FC = () => {
   const { state, dispatch } = useContext(AppContext);
@@ -68,7 +67,10 @@ const VideoList: React.FC = () => {
           {/* {rowNum === state.videoDetail.rowNum && state.videoDetail.isOpen && ( */}
           {rowNum === state.videoDetail.rowNum && (
             <WrapVideoDetailGrid item xs={12}>
-              <VideoDetail video={state.videoDetail.video} isOpen={state.videoDetail.isOpen}/>
+              <VideoDetail
+                video={state.videoDetail.video}
+                isOpen={state.videoDetail.isOpen}
+              />
             </WrapVideoDetailGrid>
           )}
         </>
